requirements.txt
django-colorfield
django-jquery
libsass
django-compressor
django-sass-processor
django-static-precompiler
